sys1/requirements.txt
amqp==2.3.2
billiard==3.5.0.3
celery==4.2.0
coverage==4.5.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.0.6
django-adminlte2==0.3.0
django-celery-beat==1.1.1
django-celery-results==1.0.1
djangorestframework==3.8.2
whitenoise==3.3.1
kombu==4.2.1
model-mommy==1.6.0
psycopg2-binary==2.7.5
python-decouple==3.1
pytz==2018.4
redis==2.10.6
six==1.11.0
static3==0.7.0
vine==1.1.4
